Higher residency in Haddock
I think there really is a slight worsening in the situation here, but
it needs someone to build a profiled compiler and take a proper look.
There isn't anything obvious in what I've done, which mainly concerns
Template Haskell and hence should not affect Haddock.

Strange, annoying, but I'm going to live with it.