Add `.mailmap` file
This improves the output of `git shortlog` and others, see

  http://git-scm.com/docs/git-shortlog

for more details.

The mapping is in part derived from the Darcs author-spelling file
which was removed some time ago via 47a018026aee9faef28ddc9b4550425dd0000ceb.

There's still a couple of unmapped `<unknown>` authors in the Git history which will
be hopefully mapped in the future as well to provide complete authorship information
throughout all of GHC's recorded development history.

Signed-off-by: Herbert Valerio Riedel <hvr@gnu.org>