Fix a bug in codegen for non-updatable selector thunks (#8817)
To evaluate most non-updatable thunks, we can jump directly to the
entry code if we know what it is.  But not for a selector thunk: these
might be updated by the garbage collector, so we have to enter the
closure with an indirect jump through its info pointer.

(cherry picked from commit b1ddec1e6d4695d71d38b59db26829d71ad784e1)