Revert ad15c2, which causes Windows seg-faults (Trac #8834)
We don't yet understand WHY commit ad15c2, which is to do with
CmmSink, causes seg-faults on Windows, but it certainly seems to.  So
reverting it is a stop-gap, but we need to un-block the 7.8 release.

Many thanks to awson for identifying the offending commit.

(cherry picked from commit a79613a75c7da0d3d225850382f0f578a07113b5)