Include EXTRA_LD_OPTS (amongst other things) when linking programs
One problem was that we weren't including $1_$2_DIST_LD_OPTS when
linking a program, which looks to be accidental: it was being defined
but not used anywhere.  This meant that setting $1_$2_EXTRA_LD_OPTS,
for example, had no effect.

This commit straightens out the handling of LD_OPTS to be consistent
with the way we handle CC_OPTS and HC_OPTS.

(cherry picked from commit 975e9cb8e7744a0750bb6c8763f628e05672643e)